Created package net.pms.exception: - Created new class InitializationException - Formatted, added JavaDocs and moved other exception implementations to the new package
