+++ src/main/java/net/pms/configuration/PmsConfiguration.java
@@ -50,6 +50,7 @@ import net.pms.encoders.Player;
 import net.pms.encoders.PlayerFactory;
 import net.pms.encoders.PlayerId;
 import net.pms.encoders.StandardPlayerId;
+import net.pms.exception.InvalidArgumentException;
 import net.pms.formats.Format;
 import net.pms.newgui.NavigationShareTab.SharedFoldersTableModel;
 import net.pms.service.PreventSleepMode;
@@ -59,7 +60,6 @@ import net.pms.util.FilePermissions;
 import net.pms.util.FileUtil;
 import net.pms.util.FileUtil.FileLocation;
 import net.pms.util.FullyPlayedAction;
-import net.pms.util.InvalidArgumentException;
 import net.pms.util.Languages;
 import net.pms.util.LogSystemInformationMode;
 import net.pms.util.PropertiesUtil;
