+++ src/main/java/net/pms/dlna/DLNAImage.java
@@ -29,12 +29,13 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import com.drew.metadata.Metadata;
 import net.pms.dlna.DLNAImageProfile.DLNAComplianceResult;
+import net.pms.exception.DLNAProfileException;
+import net.pms.exception.ParseException;
 import net.pms.image.Image;
 import net.pms.image.ImageFormat;
 import net.pms.image.ImageInfo;
 import net.pms.image.ImagesUtil;
 import net.pms.image.ImagesUtil.ScaleType;
-import net.pms.util.ParseException;
 
 /**
  * This class is simply a byte array for holding an {@link ImageIO} supported
