+++ src/main/java/net/pms/dlna/DLNAMediaInfo.java
@@ -32,6 +32,7 @@ import net.pms.configuration.PmsConfiguration;
 import net.pms.configuration.RendererConfiguration;
 import net.pms.encoders.PlayerFactory;
 import net.pms.encoders.StandardPlayerId;
+import net.pms.exception.UnknownFormatException;
 import net.pms.formats.AudioAsVideo;
 import net.pms.formats.Format;
 import net.pms.formats.Format.Identifier;
@@ -53,7 +54,6 @@ import net.pms.util.MpegUtil;
 import net.pms.util.ProcessUtil;
 import net.pms.util.Rational;
 import net.pms.util.StringUtil;
-import net.pms.util.UnknownFormatException;
 import static net.pms.util.StringUtil.*;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
