+++ src/main/java/net/pms/dlna/LibMediaInfoParser.java
@@ -13,6 +13,7 @@ import net.pms.configuration.FormatConfiguration;
 import net.pms.configuration.RendererConfiguration;
 import net.pms.dlna.DLNAMediaInfo.RateMode;
 import net.pms.dlna.MediaInfo.StreamType;
+import net.pms.exception.UnknownFormatException;
 import net.pms.formats.Format;
 import net.pms.formats.v2.SubtitleType;
 import net.pms.image.ImageFormat;
@@ -20,7 +21,6 @@ import net.pms.image.ImagesUtil;
 import net.pms.image.ImagesUtil.ScaleType;
 import net.pms.util.FileUtil;
 import net.pms.util.StringUtil;
-import net.pms.util.UnknownFormatException;
 import org.apache.commons.codec.binary.Base64;
 import static org.apache.commons.lang3.StringUtils.*;
 import org.slf4j.Logger;
