+++ src/main/java/net/pms/dlna/protocolinfo/ProtocolInfo.java
@@ -37,7 +37,7 @@ import org.fourthline.cling.support.model.dlna.DLNAProfiles;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import net.pms.dlna.protocolinfo.ProtocolInfoAttributeName.KnownProtocolInfoAttributeName;
-import net.pms.util.ParseException;
+import net.pms.exception.ParseException;
 
 
 /**
