+++ src/main/java/net/pms/formats/image/RAW.java
@@ -13,11 +13,11 @@ import net.pms.dlna.DLNAThumbnail;
 import net.pms.dlna.InputFile;
 import net.pms.encoders.PlayerFactory;
 import net.pms.encoders.DCRaw;
+import net.pms.exception.ParseException;
 import net.pms.image.ImageFormat;
 import net.pms.image.ImageInfo;
 import net.pms.image.ImagesUtil;
 import net.pms.image.ImagesUtil.ScaleType;
-import net.pms.util.ParseException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import com.drew.imaging.FileType;
