+++ src/main/java/net/pms/image/ExifInfo.java
@@ -5,8 +5,8 @@ import java.awt.image.ColorModel;
 import java.util.Arrays;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import net.pms.util.InvalidStateException;
-import net.pms.util.ParseException;
+import net.pms.exception.InvalidStateException;
+import net.pms.exception.ParseException;
 import com.drew.metadata.Directory;
 import com.drew.metadata.Metadata;
 import com.drew.metadata.exif.ExifDirectoryBase;
