+++ src/main/java/net/pms/image/ImageInfo.java
@@ -31,9 +31,9 @@ import net.pms.dlna.DLNAImageInputStream;
 import net.pms.dlna.DLNAResource;
 import net.pms.dlna.DLNAThumbnail;
 import net.pms.dlna.DLNAThumbnailInputStream;
+import net.pms.exception.InvalidStateException;
+import net.pms.exception.ParseException;
 import net.pms.image.ExifInfo.ExifParseInfo;
-import net.pms.util.InvalidStateException;
-import net.pms.util.ParseException;
 
 /**
  * This holds information about a given image, and is used as a standard image
