+++ src/main/java/net/pms/image/ImagesUtil.java
@@ -25,13 +25,13 @@ import net.pms.dlna.DLNAImageProfile;
 import net.pms.dlna.DLNAImageProfile.DLNAComplianceResult;
 import net.pms.dlna.DLNAMediaInfo;
 import net.pms.dlna.DLNAThumbnail;
+import net.pms.exception.InvalidStateException;
+import net.pms.exception.ParseException;
+import net.pms.exception.UnknownFormatException;
 import net.pms.image.ImageIOTools.ImageReaderResult;
 import net.pms.image.thumbnailator.ExifFilterUtils;
 import net.pms.util.BufferedImageType;
-import net.pms.util.InvalidStateException;
-import net.pms.util.ParseException;
 import net.pms.util.ResettableInputStream;
-import net.pms.util.UnknownFormatException;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
