+++ src/main/java/net/pms/image/JPEGInfo.java
@@ -6,7 +6,7 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
-import net.pms.util.ParseException;
+import net.pms.exception.ParseException;
 import com.drew.metadata.Directory;
 import com.drew.metadata.Metadata;
 import com.drew.metadata.jfif.JfifDirectory;
