+++ src/main/java/net/pms/newgui/SelectRenderers.java
@@ -32,8 +32,8 @@ import net.pms.Messages;
 import net.pms.PMS;
 import net.pms.configuration.PmsConfiguration;
 import net.pms.configuration.RendererConfiguration;
-import net.pms.newgui.components.IllegalChildException;
 import net.pms.newgui.components.SearchableMutableTreeNode;
+import net.pms.newgui.components.SearchableMutableTreeNode.IllegalChildException;
 import net.pms.util.tree.CheckTreeManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
