--- src/main/java/net/pms/newgui/components/IllegalChildException.java
@@ -1,45 +1,0 @@
-/*
- * Digital Media Server, for streaming digital media to UPnP AV or DLNA
- * compatible devices based on PS3 Media Server and Universal Media Server.
- * Copyright (C) 2016 Digital Media Server developers.
- *
- * This program is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any later
- * version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
- * details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program. If not, see http://www.gnu.org/licenses/.
- */
-package net.pms.newgui.components;
-
-public class IllegalChildException extends Exception {
-
-	private static final long serialVersionUID = 1152260088011461750L;
-
-	public IllegalChildException() {
-	}
-
-	public IllegalChildException(String message) {
-		super(message);
-	}
-
-	public IllegalChildException(Throwable cause) {
-		super(cause);
-	}
-
-	public IllegalChildException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-	public IllegalChildException(String message, Throwable cause,
-			boolean enableSuppression, boolean writableStackTrace) {
-		super(message, cause, enableSuppression, writableStackTrace);
-	}
-
-}
