+++ src/main/java/net/pms/service/SleepManager.java
@@ -25,8 +25,8 @@ import com.sun.jna.Platform;
 import net.pms.Messages;
 import net.pms.PMS;
 import net.pms.configuration.PmsConfiguration;
+import net.pms.exception.IOKitException;
 import net.pms.io.WinUtils.Kernel32;
-import net.pms.util.jna.macos.iokit.IOKitException;
 import net.pms.util.jna.macos.iokit.IOKitUtils;
 
 /**
